requirements.txt
torch_geometric==2.4.0
torchmetrics==0.11.0
cosine-annealing-warmup @ git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup@12d03c07553aedd3d9e9155e2b3e31ce8c64081a
easydict==1.10
einops==0.7.0
einops-exts==0.0.4
hydra-core==1.3.2
imageio==2.31.0
imageio-ffmpeg==0.4.2
lightning @ git+https://github.com/Lightning-AI/lightning@eb8b314f7cbb5e87e0d149f95f45f1409a7715c2
lightning-utilities==0.8.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
moviepy==1.0.3
numpy==1.24.4
omegaconf==2.3.0
Pillow==10.0.0
pymeshlab==2022.2.post2
pytorch-lightning==2.0.3
scikit-image==0.21.0
scikit-learn==1.2.1
scipy==1.10.1
tqdm==4.64.1
transformers==4.33.0
trimesh==3.18.0
vector-quantize-pytorch==1.8.1
wandb==0.13.4
randomname
